Refactor server files for easier testing and better modularity
