+++ server/server.test.js
@@ -1,0 +1,7 @@
+describe("Server", () => {
+  it('should return index.html from "/" endpoint', async (done) => {
+
+  });
+
+  it('requests to "/api/fec2/hrnyc/" should return data from Atlier API', () => {});
+});
